+++ src/resolvers/add_liquidity_provider.py
@@ -4,8 +4,8 @@ import time
 import requests
 
 from src.mutations.add_liquidity_provider import add_liquidity_provider_db_mutation
-from src.mutations.reset_liquidity_provider_db import reset_liquidity_provider_db_mutation
-from src.mutations.update_liquidity_provider_db import update_liquidity_provider_db_mutation
+from src.mutations.reset_liquidity_provider import reset_liquidity_provider_db_mutation
+from src.mutations.update_liquidity_provider import update_liquidity_provider_db_mutation
 from src.services.config import config_service
 from src.utils.setup_logger import setup_logger_util
